requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
python-dotenv==1.0.1
openmeteo_requests==1.3.0
google-generativeai==0.8.4
python-dotenv==1.0.1